day 2 a

+++ day_1_b/src/main.rs
@@ -4,14 +4,14 @@ fn main() {
     let file = std::fs::File::open("./input.txt");
     let lines = io::BufReader::new(file.unwrap()).lines();
 
-    let mut first_column: Vec<i32> = Vec::new();
-    let mut second_column: Vec<i32> = Vec::new();
+    let mut first_column: Vec<usize> = Vec::new();
+    let mut second_column: Vec<usize> = Vec::new();
 
     for line in lines.map_while(Result::ok) {
-        let first = &line[..5].parse::<i32>().unwrap();
+        let first = &line[..5].parse::<usize>().unwrap();
         first_column.push(*first);
 
-        let second = &line[8..13].parse::<i32>().unwrap();
+        let second = &line[8..13].parse::<usize>().unwrap();
         second_column.push(*second);
     }
 
@@ -22,7 +22,7 @@ fn main() {
 
     for n in first_column {
         let occurances = second_column.iter().filter(|&x| *x == n).count();
-        total_occurances += occurances as i32 * n;
+        total_occurances += occurances * n;
     }
 
     println!("{}", total_occurances);
